Don't disable sysstat cronjob
Although CCE-80273-6 requires us to disable the "sysstat" service, the
suggested remediation leaves the default cronjob enabled. This cronjob
invokes the "system activity accounting tool" every ten minutes as well
as generating "a daily summary of process accounting" every night.

Originally, I disabled this cronjob (by setting the permissions on the
/etc/cron.d/sysstat file to 0000). This, however, causes the check in
CCE-27209-6 ("Verify and Correct File Permissions with RPM") to fail.

This change removes the disabling of the cronjob, leaving it untouched.