Reconfigure periodic AIDE check to appease the gods
In commit 45e2998, we considered "once a week, on Sundays" to be both
"periodic" and acceptable. The gods, however, have deemed this to not be
appropriate and instead require that the check run exactly once per day
at exactly 4:05 AM if we want to receive a "pass".

Here, we reconfigure the cronjob to appease them and resign ourselves to
change it to weekly once we've passed the initial system scans.