CCE-27455-5: sshd_use_approved_macs
nb: In order to actually achieve a "pass", we have to include
"hmac-sha1" in the list of permitted MACs -- even though the use of SHA1
has been deprecated and the only SSHv2 hash algorithm that meets this
("Use Only FIPS Approved MACs") requirement is SHA2.

We'll "fix" this by removing "hmac-sha1" after our scans are performed.