Last commit: final scan reports (commit fcf82d8)
This is the last commit; this project is "complete".

If you look at these scan reports, you'll notice that there are still
failures being reported. Well, yes and no.

We certainly haven't achieved a 100% pass rate for all controls; that
much is clear. We did, however, get most of the way there.

The remaining controls (i.e. those which have not achieved "pass") can
be divided up into a couple of categories: false positive, outright
failures, and "technically it's a fail -- but not really".

False positives are exactly that: we are in fact in compliance with the
requirements but a failure is still being reported (see, for example,
the "Add {nodev,noexec,nosuid} Option to Removable Media Partitions"
controls. These were failing from the start and will fail even on a
system with *NO* "removable media partitions". I even added an explicit
entry to /etc/fstab for my CD/DVD drive -- with the required mount
options set -- and yet the checks still fail. The "Set Daemon Umask"
control has also failed from day one, even though we have set it
correctly. I gave up and said "fuck it" and moved on.

Outright failures are those controls which I have simply decided not to
address. There are only one or two of these, such as "Enable Smart Card
Login" in the PCI-DSS profile. Most enterprise environments (other than
DoD) aren't using smart cards so this isn't a big issue. There's no
point in wasting any time on these.

Last, we have "technically it's a fail -- but not really". Strictly
speaking, these are, indeed, failures; we do not meet the requirements
exactly. A very good example of this is the "Set Lockout Time For Failed
Password Attempts" control. Several of the security profiles include
this control but most of them specify different required values for the
"lockout time": C2S/NIST: 900 seconds; PCI-DSS: 1800 seconds; OSPP/STIG:
604800 seconds (one week!). If the setting doesn't match *exactly* -- it
doesn't matter if you *exceed* it -- you fail. Another good example is
the login banner that many controls require. If you don't use the login
banner that OpenSCAP says you must, you fail -- nevermind the absurdity
of declaring that a host is a "U.S. Government (USG) Information System
(IS) that is provided for USG-authorized use only" when it certainly
isn't. We all know it's the login banner that makes a host secure,
right?  *rolls eyes*

Thus, this is probably "as good as it's gonna get" and, for that reason,
you can consider this project complete.

I am going to start from scratch working on some new kickstart files,
however. There will be a kickstart file for servers that will likely be
very similar to this one (minus some of the "over the top" settings that
NIST/STIG/USGCB require). The goal is ensuring that a newly installed
system is "safe, but sensible". In addition, I'll be putting out another
kickstart file that is intended to set up a workstation environment that
is also "safe, but sensible".

Oh, last thing. I'll be focusing on CentOS instead of RHEL. The upcoming
files will almost certainly work on RHEL also (with some slight changes)
but they'll be designed with CentOS in mind -- and leaving out all of
the "Red Hat-isms".

-bofh